+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Added
 
-* `--mode` option to `judge`, `restore`, and `submit`.
+* `--mode` option to `judge`, `retrieve submissions`, and `submit`.
 * `--release` option to `judge`, which equivalent to `--mode release`.
 * `--debug` option to `submit`, which equivalent to `--mode debug`.
 * Add `.command_line_arguments`, `.config`, `.base_directory`, and `.target` to `$SNOWCHAINS_RESULT` s.
@@ -10,6 +10,9 @@
 
 ### Changed
 
+* Rename `download` command to `retrive testcases` (`download` remains as an alias)
+* Rename `restore` command to `retrive submissions`
+* Rename `list-langs` command to `retrive languages`
 * Make `env` field more configurable.
 * Now it prints most of the information to stderr, not stdout.
 * Move `service`, `contest` and `language` fields of `snowchains.toml` to another file.
@@ -23,8 +26,8 @@
 ### Fixed
 
 * Fix the condition of enabling ANSI color when `--color auto`.
-* Fix a bug where `list-langs` displays &lt;number of problems&gt; copies of a list.
-* Prevent `list-langs` from displaying unavailable languages when `problem` is specified.
+* Fix a bug where `retrive languages` displays &lt;number of problems&gt; copies of a list.
+* Prevent `retrieve languages` from displaying unavailable languages when `problem` is specified.
 * Change the default value of `jobs` from `1` to the number of CPUs.
 
 ## [0.2.0] - 2019-03-03